moved methods in a dedicated file
