+++ tests/test_scheduler.py
@@ -19,11 +19,8 @@ from streamflow.core.deployment import (
 from streamflow.core.scheduling import AvailableLocation, Hardware
 from streamflow.core.workflow import Job, Status
 from streamflow.deployment.connector import LocalConnector
-from tests.conftest import (
-    get_deployment_config,
-    get_docker_deployment_config,
-    get_service,
-)
+from tests.utils.get_instances import get_docker_deployment_config
+from tests.conftest import get_service, get_deployment_config
 
 
 @pytest_asyncio.fixture(scope="module")
